requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
openai>=1.51.0
waitress==2.1.2
requests==2.32.3
